=> Polished the Student related REST api. => Test container added to the PostgreSql database.
