+++ src/main/java/com/assignment/college_management_system/controllers/AdmissionRecordController.java
@@ -3,6 +3,7 @@ package com.assignment.college_management_system.controllers;
 import com.assignment.college_management_system.dtos.AdmissionRecordDTO;
 import com.assignment.college_management_system.services.AdmissionRecordService;
 import com.assignment.college_management_system.services.StudentService;
+import com.assignment.college_management_system.services.StudentServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,7 +15,7 @@ public class AdmissionRecordController {
     AdmissionRecordService admissionRecordService;
     StudentService studentService;
 
-    public AdmissionRecordController(AdmissionRecordService admissionRecordService, StudentService studentService) {
+    public AdmissionRecordController(AdmissionRecordService admissionRecordService, StudentServiceImpl studentService) {
         this.admissionRecordService = admissionRecordService;
         this.studentService = studentService;
     }
